Rename package to pinotquickstart
